Correct image orientation in File Chooser Preview
This makes the file chooser preview account for the embedded image orientation
in the selected file. I had to move away from the asynchronous API for this,
but since we’re only reading a single image at a time, I thought it was
justified for this added functionality. I also noted that AsyncImage was
removed from Granite, so that was probably a deprecated API anyways.